app/requirements.txt
cassandra-driver==3.28.0
Faker==19.3.0
pandas==2.0.3
pymongo==4.4.1
